+++ src/Dialog/Dialog.test.js
@@ -1,5 +1,5 @@
 import { Dialog } from '../..';
-import { events } from '../lib/events';
+import { events } from '../../.jest/events';
 
 const {
   click,
